requirements.txt
ConfigParser
RPi.GPIO
bitstring
curlify
evdev
huey
mpyg321
ntplib
pexpect
pycryptodome
pyserial
python-json-logger
requests
smbus
supervisor
transitions
